move all screenOptions to separate file
